+++ ui/src/features/Dialog/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./Content";
+export * from "./ContentHeader";
+export * from "./ContentSection";
