+++ ui/src/mock_data/workflowData.ts
@@ -1,10 +1,13 @@
+import { initialEdges, initialNodes } from "./nodeEdgeData";
+
 export function generateWorkflows(count: number) {
   const workflows = [];
   for (let i = 0; i < count; i++) {
     workflows.push({
       id: i.toString(),
       name: `My Workflow ${i + 1}`,
-      // ...nodes
+      nodes: initialNodes,
+      edges: initialEdges,
     });
   }
   return workflows;
